Refactor: MutationErrorAlertBar処理をトップ階層に移動(#71)
ユーザ削除APIカスタムフックで、使用しないonError処理を削除
