+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=0cfac6d2409bfc66230b",
+    "/js/app.js": "/js/app.js?id=775dfb314d2f29a31545",
     "/css/app.css": "/css/app.css?id=af2dc2a82d7e9d5e698b"
 }
