+++ src/resources/ts/app.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, useCallback } from 'react';
 import ReactDOM from 'react-dom';
 import {
   BrowserRouter as Router,
@@ -14,7 +14,9 @@ import Login from './containers/pages/Login';
 import Memo from './containers/pages/Memo';
 import Account from './containers/pages/Account';
 import Loding from './components/pages/Loding';
+import MutationErrorAlertBar from './components/molecules/MutationErrorAlertBar';
 import { useGetUserQuery, useCurrentUser } from './hooks/user';
+import { useMutationErrorQuery } from './hooks/util';
 
 /**
  * First we will load all of this project's JavaScript dependencies which
@@ -85,6 +87,19 @@ const App: FC = () => {
     },
   });
 
+  const { data: error } = useMutationErrorQuery();
+
+  const handleErrorBarClose = useCallback(
+    (event?: React.SyntheticEvent, reason?: string) => {
+      if (reason === 'clickaway') {
+        return;
+      }
+
+      queryClient.resetQueries('error');
+    },
+    [queryClient]
+  );
+
   if (isLoading) {
     return <Loding />;
   }
@@ -96,9 +111,17 @@ const App: FC = () => {
       </UnAuthRoute>
       <AuthRoute path="/settings/account">
         <Account />
+        <MutationErrorAlertBar
+          error={error}
+          handleErrorBarClose={handleErrorBarClose}
+        />
       </AuthRoute>
       <AuthRoute path="/:memoId?">
         <Memo />
+        <MutationErrorAlertBar
+          error={error}
+          handleErrorBarClose={handleErrorBarClose}
+        />
       </AuthRoute>
     </Switch>
   );
