+++ src/resources/ts/components/pages/Memo.tsx
@@ -7,8 +7,6 @@ import { useTheme } from '@material-ui/core/styles';
 import Header from '../../containers/organisms/Header';
 import MemoList from '../../containers/organisms/MemoList';
 import MemoDetail from '../../containers/organisms/MemoDetail';
-import MutationErrorAlertBar from '../molecules/MutationErrorAlertBar';
-import { MutationError } from '../../models/MutationError';
 
 type BodyProps = {
   memoId: string;
@@ -39,11 +37,9 @@ const MemoBody: FC<BodyProps> = ({ memoId }) => {
 
 type Props = {
   memoId: string;
-  error?: MutationError;
-  handleErrorBarClose: (event?: React.SyntheticEvent, reason?: string) => void;
 };
 
-const Memo: FC<Props> = ({ memoId, error, handleErrorBarClose }) => {
+const Memo: FC<Props> = ({ memoId }) => {
   const theme = useTheme();
   const matches = useMediaQuery(theme.breakpoints.up('sm'));
 
@@ -57,10 +53,6 @@ const Memo: FC<Props> = ({ memoId, error, handleErrorBarClose }) => {
       ) : (
         <MemoBody memoId={memoId} />
       )}
-      <MutationErrorAlertBar
-        error={error}
-        handleErrorBarClose={handleErrorBarClose}
-      />
     </>
   );
 };
