+++ src/resources/ts/hooks/user/useDeleteUserMutation.ts
@@ -1,6 +1,5 @@
 import { useQueryClient, UseMutationResult, useMutation } from 'react-query';
 import axios, { AxiosError } from 'axios';
-import { MutationError } from '../../models/MutationError';
 
 const deleteUser = async (): Promise<void> => {
   await axios.delete('/api/users/me');
@@ -20,13 +19,6 @@ const useDeleteUserMutation = (): UseMutationResult<
         predicate: (query) => query.queryKey !== 'user',
       });
       queryClient.resetQueries('user');
-    },
-    onError: (error) => {
-      const mutationError: MutationError = {
-        statusCode: error.response?.status,
-        errorMessage: 'ユーザアカウントの削除に失敗しました。',
-      };
-      queryClient.setQueryData('error', mutationError);
     },
   });
 };
